Add birthday data to birthdays.csv and send birthday letters
This commit adds data to the birthdays.csv file and implements functionality to send birthday letters. It checks if the current date matches any birthdays in the csv file and generates a random letter template for the birthday person. The letter is then customized with the person's name and sent to their email address.